Add hyperlink for the regular issue tracker
Add hyperlink for the regular issue tracker that points to `https://github.com/GrapheneOS/os-issue-tracker/issues`.

The documentation refers to the link by the name `global OS issue tracker` (not `regular issue tracker`) and suggest using it `for GrapheneOS sub-projects`. If this is not the right place to report Messaging issues, please provide the correct hyperlink.